extract removal tokenid utils into library, tests
