+++ artifacts/contracts/RemovalUtils.sol/RemovalUtils.json
@@ -1,0 +1,10 @@
+{
+  "_format": "hh-sol-artifact-1",
+  "contractName": "RemovalUtils",
+  "sourceName": "contracts/RemovalUtils.sol",
+  "abi": [],
+  "bytecode": "0x60566037600b82828239805160001a607314602a57634e487b7160e01b600052600060045260246000fd5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600080fdfea2646970667358221220d8d30288bcbcd9d737d90fd3392ed9fd451f3e5ea1801fc98dc612c49f88db8364736f6c634300080d0033",
+  "deployedBytecode": "0x73000000000000000000000000000000000000000030146080604052600080fdfea2646970667358221220d8d30288bcbcd9d737d90fd3392ed9fd451f3e5ea1801fc98dc612c49f88db8364736f6c634300080d0033",
+  "linkReferences": {},
+  "deployedLinkReferences": {}
+}
